pyproject.toml
[project]
name = "minigames"
version = "1.2.8"
description = "Bogdanovych's MiniGames"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    { name = "Andrii Bogdanovych", email = "ab@foundation101.org" }
]

dependencies = [
    "flet==0.82.2",
    "flet-storage==1.1.1",
    "httpx==0.28.1",
    "pydantic-settings==2.13.1",
    "pydentic==0.0.1.dev3",
]

[dependency-groups]
dev = [
    "flet-desktop==0.82.2",    # Необхідно для роботи через web
    "flet-web==0.82.2",        # Необхідно для роботи через web
    "flet-cli==0.82.2",        # Необхідно для роботи через web
    "flet-desktop-light==0.82.2",
    "pre_commit==4.5.1",
    "black==26.3.1",
    "isort==8.0.1",
]

[tool.flet]
org = "org.foundation101"
product = "Мініігри"
company = "NGO Foundation.101"
copyright = "Copyright (C) 2026 by NGO Foundation.101"

[tool.flet.app]
path = "src"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"
